requirements.txt
numpy
pillow
tqdm
rasterio
pandas
torch
transformers
datasets
huggingface_hub